Feature/Cart now working great, only left for conecting it with the data base
