+++ src/App.js
@@ -46,7 +46,7 @@ function App() {
 
       
     }
-  }, [dispatch, presets]);
+  }, [dispatch, presets, URL]);
   // ----------------------------------------------------------------------------------------
 
   // ----------------------------------------------------------------------------------------
